pip freeze requirements.txt
streamlit
pandas
google-generativeai
PyPDF2
python-dotenv
docx2txt
re
time
